* aoiro 0.6 金額の桁数を拡張し999億（99,999,999,999）まで扱えるようにしました。 内部処理としてはlongを使用しているため18桁まで処理できますが、12桁を超えると帳簿・決算書の一部で表示枠の幅が足りずに折り返しが発生します。

+++ src/main/java/net/osdn/aoiro/model/Creditor.java
@@ -5,7 +5,7 @@ package net.osdn.aoiro.model;
  */
 public class Creditor extends Account {
 	
-	public Creditor(AccountTitle accountTitle, int amount) {
+	public Creditor(AccountTitle accountTitle, long amount) {
 		setAccountTitle(accountTitle);
 		setAmount(amount);
 	}
